- Moved smells/ to best-practice/. - Moved deadcode/ to correctness/. - Moved JWT token checks into folders that represent the library they are checking: jose, jsonwebtoken, and passport-jwt - Moved git check to security/finding/
