+++ javascript/correctness/useless-assign.js
