+++ javascript/correctness/useless-assign.yaml
