+++ javascript/correctness/useless-eqeq.yaml
