+++ javascript/jose/security/finding/jwt-exposed-credentials.yaml
@@ -1,0 +1,220 @@
+rules:
+- id: jwt-exposed-credentials
+  languages:
+  - javascript
+  metadata:
+  - cwe: 'CWE-522: Insufficiently Protected Credentials'
+  message: |
+    Password is exposed through JWT token payload. See
+    CWE-522: Insufficiently Protected Credentials (https://cwe.mitre.org/data/definitions/522.html)
+  patterns:
+  - pattern-either:
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        var $T = JWT.sign({password:...},...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        $T = JWT.sign({password:...},...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        var $P = {password:...};
+        ...
+        var $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        var $P = {password:...};
+        ...
+        $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        $P = {password:...};
+        ...
+        var $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        $P = {password:...};
+        ...
+        $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        $P.password = ...;
+        ...
+        var $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        $P.password = ...;
+        ...
+        $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        var $P = Object.assign(...,{password:...},...);
+        ...
+        var $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        var $P = Object.assign(...,{password:...},...);
+        ...
+        $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        $P = Object.assign(...,{password:...},...);
+        ...
+        var $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        $P = Object.assign(...,{password:...},...);
+        ...
+        $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        var $T = JWT.sign(Object.assign(...,{password:...},...),...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        $T = JWT.sign(Object.assign(...,{password:...},...),...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        var $T = JWT.sign({$U:{password:...}},...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        $T = JWT.sign({$U:{password:...}},...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        var $P = {$U:{password:...}};
+        ...
+        var $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        var $P = {$U:{password:...}};
+        ...
+        $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        $P = {$U:{password:...}};
+        ...
+        var $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        $P = {$U:{password:...}};
+        ...
+        $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        $P.$U.password = ...;
+        ...
+        var $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        $P.$U.password = ...;
+        ...
+        $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        var $P = Object.assign(...,{$U:{password:...}},...);
+        ...
+        var $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        var $P = Object.assign(...,{$U:{password:...}},...);
+        ...
+        $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        $P = Object.assign(...,{$U:{password:...}},...);
+        ...
+        var $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        $P = Object.assign(...,{$U:{password:...}},...);
+        ...
+        $T = JWT.sign($P,...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        var $T = JWT.sign(Object.assign(...,{$U:{password:...}},...),...);
+    - pattern: |
+        var $JOSE = require("jose");
+        ...
+        var { JWT } = $JOSE;
+        ...
+        $T = JWT.sign(Object.assign(...,{$U:{password:...}},...),...);
+  severity: ERROR
