+++ javascript/jose/security/finding/jwt-hardcode.yaml
@@ -3,46 +3,6 @@ rules:
     patterns:
       - pattern-either:
           - pattern: |
-              var $JWT = require("jsonwebtoken");
-              ...
-              var $T = $JWT.sign($P, "...", ...);
-          - pattern: |
-              var $JWT = require("jsonwebtoken");
-              ...
-              var $T = $JWT.verify($P, "...", ...);
-          - pattern: |
-              var $JWT = require("jsonwebtoken");
-              ...
-              $JWT.sign($P, "...", ...);
-          - pattern: |
-              var $JWT = require("jsonwebtoken");
-              ...
-              $JWT.verify($P, "...", ...);
-          - pattern: |
-              var $JWT = require("jsonwebtoken");
-              ...
-              var $SECRET = "...";
-              ...
-              var $T = $JWT.sign($P, $SECRET, ...);
-          - pattern: |
-              var $JWT = require("jsonwebtoken");
-              ...
-              var $SECRET = "...";
-              ...
-              var $T = $JWT.verify($P, $SECRET, ...);
-          - pattern: |
-              var $JWT = require("jsonwebtoken");
-              ...
-              var $SECRET = "...";
-              ...
-              $JWT.sign($P, $SECRET, ...);
-          - pattern: |
-              var $JWT = require("jsonwebtoken");
-              ...
-              var $SECRET = "...";
-              ...
-              $JWT.verify($P, $SECRET, ...);
-          - pattern: |
               var $JOSE = require("jose");
               ...
               var { JWT } = $JOSE;
