+++ javascript/jose/security/finding/jwt-none-alg.js
@@ -1,0 +1,4 @@
+// ruleid: jwt-none-alg
+const jose = require("jose");
+const { JWK, JWT } = jose;
+const token = JWT.verify('token-here', JWK.None);
