+++ javascript/jose/security/finding/jwt-none-alg.yaml
@@ -5,19 +5,6 @@ rules:
     severity: ERROR
     patterns:
       - pattern-either:
-          # jsonwebtoken
-          - pattern: |
-              var $JWT = require("jsonwebtoken");
-              ...
-              var $T = $JWT.verify($P, $X, {algorithms:[...,'none',...]},...);
-          - pattern: |
-              var $JWT = require("jsonwebtoken");
-              ...
-              $T = $JWT.verify($P, $X, {algorithms:[...,'none',...]},...);
-          - pattern: |
-              var $JWT = require("jsonwebtoken");
-              ...
-              $JWT.verify($P, $X, {algorithms:[...,'none',...]},...);
           - pattern: |
               var $JOSE = require("jose");
               ...
