+++ javascript/jsonwebtoken/security/audit/example1.js
