+++ javascript/jsonwebtoken/security/finding/example1.js
