+++ javascript/jsonwebtoken/security/finding/example2.js
