+++ javascript/jsonwebtoken/security/finding/example3.js
