+++ javascript/jsonwebtoken/security/finding/example4.js
