+++ javascript/jsonwebtoken/security/finding/jwt-none-alg.js
@@ -1,8 +1,3 @@
-// ruleid: jwt-none-alg
-const jose = require("jose");
-const { JWK, JWT } = jose;
-const token = JWT.verify('token-here', JWK.None);
-
 function verifyJwt() {
 	// ruleid: jwt-none-alg
 	let jwt = require("jsonwebtoken");
