+++ javascript/passport-jwt/security/finding/docs-examples.js
