+++ javascript/passport-jwt/security/finding/example-1.js
