+++ javascript/passport-jwt/security/finding/example-2.js
