+++ javascript/passport-jwt/security/finding/example-3.js
