+++ javascript/passport-jwt/security/finding/example-4.js
