+++ javascript/passport-jwt/security/finding/generate-passport-rules.py
