+++ javascript/passport-jwt/security/finding/passport-hardcode.yaml
