+++ javascript/security/finding/spawn-git-clone.js
