+++ javascript/security/finding/spawn-git-clone.yaml
