+++ python/boto3/security/finding/hardcoded-token.py
