+++ python/requests/security/finding/no-auth-over-http.py
